fixed incorrect property name, $(GetCopyToOutputDirectoryItemsDependsOn) instead of $(GetCopyToOutputDirectoryItems)
The intention here is to add to the existing GetCopyToOutputDirectoryItemsDependsOn property, defined in Microsoft.Common.Targets. Instead what was happening is that it was being completely overwitten with the _RazorGetCopyToOutputDirectoryItems entry, given that $(GetCopyToOutputDirectoryItems) doesn't exist as a property and must have been a typo.

One effect of this is that if a project references an RCL that has a non-cshtml file marked as Content (copy to output), this file will only be copied to the consuming projects output directory during a full rebuild, or if the RCL is out-of-date and needs rebuilding.